Update and rename vsearch4web.py to view.py
