--- nfu_newmedia-master/webapp_zh/vsearch4web.py
@@ -1,49 +1,0 @@
-from flask import Flask, render_template, request, escape
-from vsearch import search4letters
-
-app = Flask(__name__)
-
-
-def log_request(req: 'flask_request', res: str) -> None:
-    """Log details of the web request and the results."""
-    with open('vsearch.log', 'a') as log:
-        print(req.form, req.remote_addr, req.user_agent, res, file=log, sep='|')
-
-        
-@app.route('/results', methods=['POST'] action='/results')
-def BMI() -> 'html':
-    height = request.form['height']
-    weight = request.form['weight']
-    
-    return render_template('results.html',
-                           the_height=height,
-                           the_weight=weight
-                           )
-
-
-@app.route('/')
-@app.route('/entry')
-def entry_page() -> 'html':
-    """Display this webapp's HTML form."""
-    return render_template('entry.html',
-                           the_title='欢迎一A健康监测系统！')
-
-
-@app.route('/viewlog')
-def view_the_log() -> 'html':
-    """Display the contents of the log file as a HTML table."""
-    contents = []
-    with open('vsearch.log') as log:
-        for line in log:
-            contents.append([])
-            for item in line.split('|'):
-                contents[-1].append(escape(item))
-    titles = ('表单内容', '访问者IP', '浏览器', '运行结果')
-    return render_template('viewlog.html',
-                           the_title='查看日志',
-                           the_row_titles=titles,
-                           the_data=contents,)
-
-
-if __name__ == '__main__':
-    app.run(debug=True)
